% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_data.R
\name{balance_data}
\alias{balance_data}
\title{Function to balance datasets}
\usage{
balance_data(
  counterfactual_method,
  treatment_variable,
  matching_variable,
  PS_estimation_object,
  missing_method,
  eff,
  ...
)
}
\arguments{
\item{counterfactual_method}{Character string indicating your chosen approach \code{"psm"}, \code{"iptw"}, or \code{"nbp"}}

\item{treatment_variable}{Character string matching column name of your treatment variable}

\item{matching_variable}{Character vector matching column name(s) of your matching variable(s)}

\item{PS_estimation_object}{Propensity score estimation object}

\item{missing_method}{Character string indicating your chosen method of dealing with missingness \code{"complete"}, \code{"mi"}, or \code{"weighting"}}

\item{eff}{}

\item{...}{}
}
\description{
Function to balance datasets
}
\examples{
estimates <- estimation_stage(
.data = DigiCAT::zp_eg,
missing_method = "complete",
model_type = "glm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
weighting_variable = NULL,
cluster_variable = NULL,
strata_variable = NULL
)

balanced_data <- balance_data(
counterfactual_method = "psm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
PS_estimation_object = estimates,
missing_method = "complete",
)
}
