% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancing.R
\name{balancing}
\alias{balancing}
\title{Balancing  function}
\usage{
balancing(cf_method, t_var, m_vars, psmodel_obj, eff, ...)
}
\arguments{
\item{cf_method}{counterfactual approach employed, one of \code{"matching"}, \code{"weighting"}, \code{"cem"}, TODO add here}

\item{t_var}{name of treatment variable in dataset (character string)}

\item{m_vars}{vector of matching variable names in dataset}

\item{psmodel_obj}{an object obtained from get_score(). this is required for matching and iptw methods, and is a list containing data, model, propensity score, and string indicating class of model. note the first 3 of these may be lists if multiple imputation has been used.}

\item{...}{additional arguments to be passed to matchit/weightit/matchthem/weightthem functions}
}
\description{
Balancing  function
}
