% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_selected_variables.R
\name{check_selected_variables}
\alias{check_selected_variables}
\title{Flag variable input errors function}
\usage{
check_selected_variables(outcome, treatment, matchvars, covars)
}
\description{
Flag variable input errors function
}
