% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_common_support.R
\name{check_support}
\alias{check_support}
\title{Function to display common support}
\usage{
check_support(estimation_model_object, missing_method, ...)
}
\arguments{
\item{estimation_model_object}{}

\item{missing_method}{}

\item{...}{}
}
\description{
Function to display common support
}
