% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_estimation_stage.R
\name{estimation_stage}
\alias{estimation_stage}
\title{Propensity score estimation}
\usage{
estimation_stage(
  .data,
  missing_method,
  model_type,
  treatment_variable,
  matching_variable,
  weighting_variable = NULL,
  cluster_variable = NULL,
  strata_variable = NULL,
  ...
)
}
\arguments{
\item{.data}{dataframe}

\item{missing_method}{Character string indicating your chosen method of dealing with missingness \code{"complete"}, \code{"mi"}, or \code{"weighting"}}

\item{model_type}{Character string indicating your chosen model for calculating propensity scores \code{"glm"}}

\item{treatment_variable}{Character string matching column name of your treatment variable}

\item{matching_variable}{Character vector/string matching column name(s) of your matching variable(s)}

\item{weighting_variable}{Character string matching column name of your weighting variable}

\item{cluster_variable}{Character string matching column name of your clustering variable}

\item{strata_variable}{Character string matching column name of your stratification variable}

\item{...}{}
}
\value{
Estimation model object
}
\description{
This function calculates the propensity score (or likelihood of belonging to the treatment group)
for each individual in the sample data, based on the matching variables provided.
}
\examples{
estimation_stage(
.data = DigiCAT::zp_eg,
missing_method = "complete",
model_type = "glm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
weighting_variable = NULL,
cluster_variable = NULL,
strata_variable = NULL
)

}
