% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categorical_variables.R
\name{get_categorical_variables}
\alias{get_categorical_variables}
\title{Get categorical column names function}
\usage{
get_categorical_variables(x)
}
\description{
Get categorical column names function
}
