% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NBP_balancing_output.R
\name{get_NBP_balancing_output}
\alias{get_NBP_balancing_output}
\title{Function to return observation table, love plot and balance table from NBP matching}
\usage{
get_NBP_balancing_output(
  estimation_model_object,
  balanced_data,
  treatment_variable,
  matching_variables,
  missingness
)
}
\arguments{
\item{estimation_model_object}{}

\item{balanced_data}{}

\item{treatment_variable}{}

\item{missingness}{}
}
\description{
Function to return observation table, love plot and balance table from NBP matching
}
