% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\title{DigiCAT function to calculate propensity scores}
\usage{
get_score(
  psmodel = "glm",
  .data,
  t_var = NULL,
  y_var = NULL,
  m_vars = NULL,
  covars = NULL,
  missing,
  ...
)
}
\arguments{
\item{psmodel}{class of model used to estimate propensity scores. character string, one of \code{"glm"} (default), \code{"gbm"}, \code{"rforest"}.}

\item{.data}{dataset including treatment variable and any matching variables}

\item{t_var}{name of treatment variable in dataset (character string)}

\item{y_var}{name of outcome variable in dataset (character string)}

\item{m_vars}{vector of matching variable names in dataset}

\item{covars}{vector of covariate names in dataset}

\item{missing}{method by which missing data is handled, one of \code{"complete"} (complete cases), \code{"mi"} (multiple imputation), or \code{"fiml"} (full information maximum likelihood). TODO fiml not implemented bcause lavaan}

\item{...}{additional arguments to be passed to model estimating functions}
}
\description{
DigiCAT function to calculate propensity scores
}
