% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missingness.R
\name{handle_missingness}
\alias{handle_missingness}
\title{Function to handle missingness in datasets}
\usage{
handle_missingness(
  .data,
  missing_method = NULL,
  counterfactual_method = NULL,
  cluster_variable = NULL,
  weighting_variable = NULL,
  strata_variable = NULL,
  ...
)
}
\arguments{
\item{.data}{}

\item{missing_method}{}

\item{...}{}

\item{design_object}{}
}
\description{
Function to handle missingness in datasets
}
