% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_model.R
\name{outcome_analysis}
\alias{outcome_analysis}
\title{DigiCAT outcome function}
\usage{
outcome_analysis(
  outcome_method,
  y_var,
  t_var,
  m_vars,
  balanced_data,
  ids = NULL,
  weights = NULL,
  strata = NULL,
  fpc = NULL,
  cf_method,
  psmod,
  ...
)
}
\arguments{
\item{outcome_method}{outcome analysis, one of \code{"unweighted linear regression"}, \code{"design-weighted linear regression"}, \code{"g-computation"}}

\item{y_var}{name of outcome variable in dataset (character string)}

\item{t_var}{name of treatment/exposure variable in dataset (character string)}

\item{m_vars}{vector of matching variables/covariates names in dataset - does not need to be matching variables}

\item{balanced_data}{an object obtained from balance()}

\item{ids}{name of variable to use in svydesign() for weighted analysis}

\item{weights}{name of variable to use in svydesign() for weighted analysis}

\item{strata}{name of variable to use in svydesign() for weighted analysis}

\item{fpc}{name of variable to use in svydesign() for weighted analysis}

\item{cf_method}{whether matching or weighting was used as determines outcome procedures later on}

\item{...}{additional arguments}
}
\description{
DigiCAT outcome function
}
