% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_upload_page.R
\name{reset_upload_page}
\alias{reset_upload_page}
\title{Functions that resets data upload page}
\usage{
reset_upload_page(
  reset_errors = FALSE,
  hide_data = FALSE,
  hide_validation = FALSE,
  parent
)
}
\arguments{
\item{reset_errors}{Do you want to reset error messages? Default is FALSE}

\item{hide_data}{Do you want to hide the data tab? Default is FALSE}

\item{hide_validation}{Do you want to hide the data tab? Default is FALSE}
}
\description{
Functions that resets data upload page
}
