% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sensitivity.R
\name{run_sensitivity}
\alias{run_sensitivity}
\title{Sensitivity Analysis}
\usage{
run_sensitivity(
  PS_estimation_object,
  balanced_data,
  missing_method,
  outcome_variable,
  sensitivity_type,
  outcome_results,
  outcome_type,
  ...
)
}
\description{
Sensitivity Analysis
}
