% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_data.R
\name{balance_data}
\alias{balance_data}
\title{Function to balance datasets}
\usage{
balance_data(
  counterfactual_method,
  treatment_variable,
  matching_variable,
  PS_estimation_object,
  missing_method,
  eff,
  ...
)
}
\arguments{
\item{counterfactual_method}{}

\item{treatment_variable}{}

\item{matching_variable}{}

\item{PS_estimation_object}{}

\item{missing_method}{}

\item{eff}{}

\item{...}{}
}
\description{
Function to balance datasets
}
