% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_R_script.R
\name{get_R_script}
\alias{get_R_script}
\title{Function to produce downloadable R script}
\usage{
get_R_script(
  data_source,
  file_path = NULL,
  categorical_variables = NULL,
  outcome_variable,
  treatment_variable,
  matching_variables,
  covariates = NULL,
  weighting_variable = NULL,
  cluster_variable = NULL,
  strata_variable = NULL,
  CF_approach,
  missing_method,
  balancing_model,
  matching_method = NULL,
  matching_ratio = NULL,
  outcome_formula,
  DigiCAT_balanced_data,
  DigiCAT_extracted_balanced_data,
  DigiCAT_fitted_model,
  DigiCAT_extracted_outcome_results
)
}
\arguments{
\item{data_source}{Source of data: \code{"own"}, or \code{"sample"}}

\item{file_path}{File path to data}

\item{categorical_variables}{Vector of categorical variable names in dataset}

\item{outcome_variable}{Name of outcome variable in dataset (character string)}

\item{treatment_variable}{Name of treatment variable in dataset (character string)}

\item{matching_variables}{Vector of matching variable names in dataset}

\item{covariates}{Vector of covariate names in dataset}

\item{CF_approach}{Name of counterfactual appraoch being taken (character string)}

\item{missing_method}{Name of method of dealing with missingness being used (character string): \code{"mi"}, or \code{"complete"}}

\item{balancing_model}{Name of balancing model being used (character string)}

\item{matching_method}{Name of balancing method being used (character string): \code{"nearest"} or \code{"optimal"}}

\item{matching_ratio}{Name of balancing ratio being used (numeric)}

\item{outcome_formula}{Name of outcome model being used (character string)}

\item{DigiCAT_balanced_data}{Balanced data object from tool}

\item{DigiCAT_extracted_balanced_data}{Balanced data object from tool extracted for outcome model run}

\item{DigiCAT_fitted_model}{Fitted outcome model from tool}

\item{DigiCAT_extracted_outcome_results}{Extracted outcome model results from tool}

\item{data}{Dataframe}
}
\description{
Function to produce downloadable R script
}
