% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation.R
\name{get_validation}
\alias{get_validation}
\title{Data validation in DigiCAT}
\usage{
get_validation(
  .data,
  treatment,
  outcome,
  matchvars,
  covars,
  survey_weight_var,
  non_response_weight,
  clustering_var,
  stratification_var
)
}
\arguments{
\item{treatment}{Name of treatment variable.}

\item{outcome}{Name of outcome variable.}

\item{matchvars}{Name of matching variables.}

\item{covars}{Name of covariates.}

\item{survey_weight_var}{Name of survey weight variable.}

\item{non_response_weight}{Is there a non-response weight variable?}

\item{clustering_var}{Name of clustering variable.}

\item{stratification_var}{Name of stratification variable.}

\item{data}{Name of upladed dataset. Default is 'inputData$rawdata'.}
}
\description{
Data validation in DigiCAT
}
