% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_validation.R
\name{get_validation}
\alias{get_validation}
\title{Data validation in DigiCAT}
\usage{
get_validation(
  .data,
  treatment,
  outcome,
  matchvars,
  covars,
  survey_weight_var,
  non_response_weight
)
}
\arguments{
\item{treatment}{Name of treatment variable.}

\item{outcome}{Name of outcome variable.}

\item{matchvars}{Name of matching variables.}

\item{data}{Name of upladed dataset. Default is 'inputData$rawdata'.}

\item{NRW_var}{Name of non-response weight variable.}
}
\description{
Data validation in DigiCAT
}
