% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missingness.R
\name{handle_missingness}
\alias{handle_missingness}
\title{Function to handle missingness in datasets}
\usage{
handle_missingness(
  .data,
  missing_method = NULL,
  design_object = NULL,
  counterfactual_method = NULL,
  ...
)
}
\arguments{
\item{.data}{}

\item{missing_method}{}

\item{design_object}{}

\item{...}{}
}
\description{
Function to handle missingness in datasets
}
