% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_analysis_stage.R
\name{outcome_analysis_stage}
\alias{outcome_analysis_stage}
\title{Counterfactual analysis outcome model}
\usage{
outcome_analysis_stage(
  balanced_data,
  counterfactual_method,
  outcome_variable,
  outcome_type,
  treatment_variable,
  matching_variable,
  covariates = NULL,
  PS_estimation_object,
  cluster_variable = NULL,
  weighting_variable = NULL,
  strata_variable = NULL,
  missing_method,
  outcome_formula,
  ...
)
}
\arguments{
\item{balanced_data}{Balanced data object}

\item{counterfactual_method}{Character string indicating your chosen approach \code{"psm"}, \code{"iptw"}, or \code{"nbp"}}

\item{outcome_variable}{Character string matching column name of your outcome variable}

\item{treatment_variable}{Character string matching column name of your treatment variable}

\item{matching_variable}{Character vector matching column name(s) of your matching variable(s)}

\item{PS_estimation_object}{Propensity score estimation object}

\item{cluster_variable}{Character string matching column name of your clustering variable}

\item{missing_method}{Character string indicating your chosen method of dealing with missingness \code{"complete"}, \code{"mi"}, or \code{"weighting"}}

\item{...}{}

\item{nonresponse_weights}{Character string matching column name of your non-response weight variable}

\item{sampling_weights}{Character string matching column name of your sampling weight variable}
}
\value{
Estimated effect of treatment on outcome
}
\description{
This function runs the outcome model for counterfactual analysis in DigiCAT.
}
\examples{
estimates <- estimation_stage(
.data = DigiCAT::zp_eg,
missing_method = "complete",
model_type = "glm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
weighting_variable = NULL,
cluster_variable = NULL,
strata_variable = NULL
)

balanced_data <- balance_data(
counterfactual_method = "psm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
PS_estimation_object = estimates,
missing_method = "complete",
)

outcome_analysis_stage(
balanced_data = balanced_data, 
counterfactual_method = "psm",
missing_method = "complete",
outcome_variable = "Anxiety_age17",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
PS_estimation_object = estimates
)
}
