% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_analysis_stage.R
\name{outcome_analysis_stage}
\alias{outcome_analysis_stage}
\title{Counterfactual analysis outcome model}
\usage{
outcome_analysis_stage(
  balanced_data,
  counterfactual_method,
  outcome_variable,
  treatment_variable,
  matching_variable,
  psmodel_obj,
  cluster_variable,
  nonresponse_weights,
  sampling_weights,
  missing_method,
  ...
)
}
\arguments{
\item{balanced_data}{}

\item{counterfactual_method}{}

\item{outcome_variable}{}

\item{treatment_variable}{}

\item{matching_variable}{}

\item{psmodel_obj}{}

\item{cluster_variable}{}

\item{nonresponse_weights}{}

\item{sampling_weights}{}

\item{missing_method}{}

\item{...}{}
}
\value{
Estimated effect of treatment on outcome
}
\description{
This function runs the outcome model for counterfactual analysis in DigiCAT.
}
\examples{
estimates <- estimation_stage(
.data = DigiCAT::zp_eg,
missing_method = "complete",
model_type = "glm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
weighting_variable = NULL,
cluster_variable = NULL,
strata_variable = NULL
)

balanced_data <- balance_data(
counterfactual_method = "psm",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
PS_estimation_object = estimates,
missing_method = "complete",
)

outcome_analysis_stage(
balanced_data = balanced_data, 
counterfactual_method = "psm",
missing_method = "complete",
outcome_variable = "Anxiety_age17",
treatment_variable = "Reading_age15",
matching_variable = names(DigiCAT::zp_eg)[-c(2:4)],
psmodel_obj = estimates
)
}
