% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_DigiCAT}
\alias{run_DigiCAT}
\title{Run DigiCAT locally}
\usage{
run_DigiCAT(enableLocal = TRUE, filePath = fs::path_home())
}
\arguments{
\item{enableLocal}{Enable the upload of local data \code{TRUE} or \code{FALSE}. Default is \code{TRUE}.}
}
\description{
Run DigiCAT locally
}
\examples{
run_DigiCAT(enableLocal = TRUE)
}
